Player implements IHandEvaluator now
