+++ objectmodel/src/game/mechanics/analysis/card/ICardAnalyzer.java
@@ -1,0 +1,11 @@
+package game.mechanics.analysis.card;
+
+import game.cardset.card.Card;
+
+/**
+ * Created by nikiforos on 10/09/15.
+ */
+public interface ICardAnalyzer {
+
+    int analyze(Card card);
+}
