--- objectmodel/src/game/mechanics/evaluation/card/ICardEvaluator.java
@@ -1,11 +1,0 @@
-package game.mechanics.evaluation.card;
-
-import game.cardset.card.Card;
-
-/**
- * Created by nikiforos on 10/09/15.
- */
-public interface ICardEvaluator {
-
-    int evaluateCard(Card card);
-}
