--- objectmodel/src/game/mechanics/evaluation/hand/DummyHandEvaluator.java
@@ -1,15 +1,0 @@
-package game.mechanics.evaluation.hand;
-
-import game.cardset.Hand;
-
-/**
- * Created by nikiforos on 07/09/15.
- */
-public class DummyHandEvaluator implements IHandEvaluator {
-
-    @Override
-    public int evaluateHand(Hand handToEvaluate) {
-        return 60;
-    }
-
-}
