+++ objectmodel/test/game/player/mock/MockClassicPlayer.java
@@ -1,21 +1,34 @@
 package game.player.mock;
 
+import game.cardset.Hand;
+import game.cardset.card.Card;
+import game.mechanics.analysis.card.FixedScaleAnalyzer;
+import game.mechanics.analysis.card.ICardAnalyzer;
+
+import java.util.Collection;
+import java.util.Iterator;
+
 /**
  * Created by nikiforos on 04/09/15.
  */
-//@RunWith(Parameterized.class)
 public class MockClassicPlayer extends MockPlayer {
 
     public MockClassicPlayer() {
         super(0.4F, 1);
     }
 
-//    @Parameterized.Parameters
-//    public static Collection initParameters() {
-//        return Arrays.asList(new Object[][]{
-//                {true},
-//                {false}
-//        });
-//    }
+    @Override
+    public int evaluateHand(Hand handToEvaluate) {
+        ICardAnalyzer cardEvaluator = new FixedScaleAnalyzer();
+        Collection<Card> cards = handToEvaluate.getCardSet();
+        int handValue = 0;
+        Iterator<Card> cardIterator = cards.iterator();
+        while (cardIterator.hasNext()) {
+            int cardValue = cardEvaluator.analyze(cardIterator.next());
+            handValue += cardValue;
+        }
 
+        int weightedHandValue = (int) (handValue / 20.0) * 3;
+        return weightedHandValue;
+    }
 }
