+++ objectmodel/test/game/player/mock/MockCowardPlayer.java
@@ -1,5 +1,7 @@
 package game.player.mock;
 
+import game.cardset.Hand;
+
 /**
  * Created by nikiforos on 04/09/15.
  */
@@ -9,4 +11,8 @@ public class MockCowardPlayer extends MockPlayer {
         super(1F, 0);
     }
 
+    @Override
+    public int evaluateHand(Hand handToEvaluate) {
+        return 60;
+    }
 }
