+++ objectmodel/test/game/player/mock/characteristic/CommonData.java
@@ -1,0 +1,14 @@
+package game.player.mock.characteristic;
+
+import game.player.mock.MockClassicPlayer;
+import game.player.mock.MockCowardPlayer;
+import game.player.mock.MockUnwaveringPlayer;
+
+/**
+ * Created by nikiforos on 10/01/16.
+ */
+class CommonData {
+
+    final static Class[] PERS_IMPL_CLASSES = {MockClassicPlayer.class, MockCowardPlayer.class, MockUnwaveringPlayer.class};
+
+}
